Before changing the particle state into f64 type

+++ processes/particles.cuh
@@ -23,14 +23,14 @@
 // Stack of particles, format data is defined as SoA
 struct ParticleStack{
     // property
-    f32* E;
-    f32* dx;
-    f32* dy;
-    f32* dz;
-    f32* px;
-    f32* py;
-    f32* pz;
-    f32* tof;
+    f64* E;
+    f64* dx;
+    f64* dy;
+    f64* dz;
+    f64* px;
+    f64* py;
+    f64* pz;
+    f64* tof;
     // PRNG
     ui32* prng_state_1;
     ui32* prng_state_2;
